Initial cluster-api provider
This commit also incorporates various commits and ideas and code from:

  https://github.com/ingvagabund/autoscaler

Jan: I took the work we did on the openshift fork and rebased (and
squashed) them all here. I also chose not to commit the cluster-api
vendor directory as there are now make rules to do that automatically
if it does not exist.

This now runs but TemplateNodeInfo() needs to be implemented next
before it will actually scale up/down.