UPSTREAM: <carry>: test/openshift/e2e: don't change replica count
We were changing the replica count when the cluster-autoscaler was
still running which meant that the test would occasionally flake.

For example, you would occasionally see the following, and note they
are only 10s apart:

```console
I0214 12:51:41.034700 1 scale_up.go:584] Scale-up: setting group kube-system/kubemark-machine-pool size to 3
I0214 12:51:51.129943 1 scale_up.go:584] Scale-up: setting group kube-system/kubemark-machine-pool size to 3
```

Between these two calls we were adjusting the replica count in an
attempt to clean up. However, when this occurred, the node count would
never drop below the initial node count recorded at the beginning of
the test and as result the test could never pass.

The fix is to not change the replica count on the machine set at all,
instead relying on the autoscaler to naturally scale down (which is
part of the test itself).