UPSTREAM: <carry>: openshift: Rework TestNodeGroupDeleteNodes
This reworks `TestNodeGroupMachineSetDeleteNodes` to use subtests. The
change is to ensure we are testing both MachineSets and
MachineDeployments in all cases we construct a nodegroup.

By switching to subtests it is easier to see those places where
machine deployments are not being tested as the the output for each
test function should show that MachineDeployment and MachineSet is
listed in the test name output.

```console
--- PASS: TestNodeGroupDeleteNodes (0.20s)
    --- PASS: TestNodeGroupDeleteNodes/MachineSet (0.10s)
    --- PASS: TestNodeGroupDeleteNodes/MachineDeployment (0.10s)
```