UPSTREAM: <carry>: openshift: drop machine annotation linkage
This changes the mapping between a node and a machine to rely on the
presence of the ProviderID rather than the "machine" annotation
currently added by the nodelink-controller.

Note: this should not merge until we update the AWS, libvirt and
kubemark actuators - they should ensure that ProviderID is set, if not
already, on the machine object.